prefect/requirements.txt
# Superseded by pyproject.toml + uv at the project root.
# Kept for reference only — use `uv sync` from the project root instead.
prefect>=2.14
google-cloud-storage>=2.13
google-cloud-bigquery>=3.13
pyarrow>=14.0
pandas>=2.1
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
